fix: resolved issue #22 #33 #23
