+++ src/modules/tui/components/bot_cogs.ts
@@ -1,4 +1,4 @@
-import { TUIComponentSchemaType } from "../types";
+import { TUIComponentSchemaType } from "../tui-types";
 
 export default {
   name: "bot_cogs",
