+++ src/modules/tui/components/general_log.ts
@@ -1,4 +1,4 @@
-import { TUIComponentSchemaType } from "../types";
+import { TUIComponentSchemaType } from "../tui-types";
 
 export default {
   name: "general_log",
